refactor(tui): split oversized state and markdown modules
