+++ .tmp-pr-split-large-files.md
@@ -1,0 +1,10 @@
+## Summary
+
+- split `src/tui/state.rs` by extracting shared TUI state types into `src/tui/state/types.rs`
+- split local-link parsing helpers out of `src/tui/markdown_render.rs` into `src/tui/markdown_render/local_links.rs`
+- keep behavior unchanged while reducing oversized source files
+
+## Validation
+
+- `cargo check`
+- verified there are no `src/**/*.rs` files over 800 lines after the split
